Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -41,11 +41,19 @@ class QuestionsController extends Controller
 
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied for edit that question.');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied for edit that question.');
+        }
+
         $question->update($request->only([
             'title',
             'body',
@@ -56,6 +64,10 @@ class QuestionsController extends Controller
 
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied for delete that question.');
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
